small fixes for repl execution during dbg
