+++ PowershellTools.Common/PowershellTools.Common.csproj
@@ -86,6 +86,7 @@
     <Compile Include="ServiceManagement\DebuggingContract\DebuggerBreakpointUpdatedEventArgs.cs" />
     <Compile Include="ServiceManagement\DebuggingContract\CallStack.cs" />
     <Compile Include="ServiceManagement\DebuggingContract\DebuggerStoppedEventArgs.cs" />
+    <Compile Include="ServiceManagement\DebuggingContract\DebuggingServiceException.cs" />
     <Compile Include="ServiceManagement\DebuggingContract\IPowershellDebuggingService.cs" />
     <Compile Include="ServiceManagement\DebuggingContract\IDebugEngineCallback.cs" />
     <Compile Include="ServiceManagement\DebuggingContract\PowershellBreakpoint.cs" />
