+++ PowershellTools.Common/ServiceManagement/DebuggingContract/IDebugEngineCallback.cs
@@ -20,7 +20,7 @@ namespace PowerShellTools.Common.ServiceManagement.DebuggingContract
         void OutputString(string output);
 
         [OperationContract(IsOneWay = true)]
-        void TerminatingException(Exception ex);
+        void TerminatingException(DebuggingServiceException ex);
 
         [OperationContract(IsOneWay = true)]
         void DebuggerFinished();
