+++ PowershellTools.HostService/PowershellTools.HostService.csproj
@@ -85,10 +85,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ServiceManagement\Debugging\PowershellDebuggingService.cs" />
-    <Compile Include="ServiceManagement\Debugging\PSHost.cs" />
+    <Compile Include="ServiceManagement\Debugging\PowershellHost.cs" />
+    <Compile Include="ServiceManagement\Debugging\PowershellServiceHost.cs" />
     <Compile Include="ServiceManagement\PowershellIntelliSenseService.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ServiceManagement\ServiceCommon.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\PowershellTools.Common\PowerShellTools.Common.csproj">
