+++ PowershellTools.HostService/ServiceManagement/Debugging/PowershellHost.cs
