+++ PowershellTools.HostService/ServiceManagement/Debugging/PowershellServiceHost.cs
@@ -1,0 +1,146 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Management.Automation.Host;
+using System.ServiceModel;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PowerShellTools.HostService.ServiceManagement.Debugging
+{
+    public partial class PowershellDebuggingService : PSHost
+    {
+        /// <summary>
+        /// The identifier of this PSHost implementation.
+        /// </summary>
+        private Guid myId = Guid.NewGuid();
+
+        /// <summary>
+        /// Gets a string that contains the name of this host implementation. 
+        /// Keep in mind that this string may be used by script writers to
+        /// identify when your host is being used.
+        /// </summary>
+        public override string Name
+        {
+            get { return "PowershellToolOutProcHost"; }
+        }
+
+        /// <summary>
+        /// This implementation always returns the GUID allocated at 
+        /// instantiation time.
+        /// </summary>
+        public override Guid InstanceId
+        {
+            get { return this.myId; }
+        }
+
+        public HostUi HostUi { get; private set; }
+
+        public override PSHostUserInterface UI
+        {
+            get { return HostUi; }
+        }
+
+        /// <summary>
+        /// Gets the version object for this application. Typically this 
+        /// should match the version resource in the application.
+        /// </summary>
+        public override Version Version
+        {
+            get { return new Version(1, 0, 0, 0); }
+        }
+
+        /// <summary>
+        /// This API Instructs the host to interrupt the currently running 
+        /// pipeline and start a new nested input loop. In this example this 
+        /// functionality is not needed so the method throws a 
+        /// NotImplementedException exception.
+        /// </summary>
+        public override void EnterNestedPrompt()
+        {
+            throw new NotImplementedException(
+                  "The method or operation is not implemented.");
+        }
+
+        /// <summary>
+        /// This API instructs the host to exit the currently running input loop. 
+        /// In this example this functionality is not needed so the method 
+        /// throws a NotImplementedException exception.
+        /// </summary>
+        public override void ExitNestedPrompt()
+        {
+            throw new NotImplementedException(
+                  "The method or operation is not implemented.");
+        }
+
+        /// <summary>
+        /// This API is called before an external application process is 
+        /// started. Typically it is used to save state so that the parent  
+        /// can restore state that has been modified by a child process (after 
+        /// the child exits). In this example this functionality is not  
+        /// needed so the method returns nothing.
+        /// </summary>
+        public override void NotifyBeginApplication()
+        {
+            return;
+        }
+
+        /// <summary>
+        /// This API is called after an external application process finishes.
+        /// Typically it is used to restore state that a child process has
+        /// altered. In this example, this functionality is not needed so  
+        /// the method returns nothing.
+        /// </summary>
+        public override void NotifyEndApplication()
+        {
+            return;
+        }
+
+        /// <summary>
+        /// Indicate to the host application that exit has
+        /// been requested. Pass the exit code that the host
+        /// application should use when exiting the process.
+        /// </summary>
+        /// <param name="exitCode">The exit code that the 
+        /// host application should use.</param>
+        public override void SetShouldExit(int exitCode)
+        {
+
+        }
+        /// <summary>
+        /// The culture information of the thread that created
+        /// this object.
+        /// </summary>
+        private CultureInfo originalCultureInfo =
+            System.Threading.Thread.CurrentThread.CurrentCulture;
+
+        /// <summary>
+        /// The UI culture information of the thread that created
+        /// this object.
+        /// </summary>
+        private CultureInfo originalUICultureInfo =
+            System.Threading.Thread.CurrentThread.CurrentUICulture;
+
+        /// <summary>
+        /// Gets the culture information to use. This implementation 
+        /// returns a snapshot of the culture information of the thread 
+        /// that created this object.
+        /// </summary>
+        public override System.Globalization.CultureInfo CurrentCulture
+        {
+            get { return this.originalCultureInfo; }
+        }
+
+        /// <summary>
+        /// Gets the UI culture information to use. This implementation 
+        /// returns a snapshot of the UI culture information of the thread 
+        /// that created this object.
+        /// </summary>
+        public override System.Globalization.CultureInfo CurrentUICulture
+        {
+            get { return this.originalUICultureInfo; }
+        }
+
+    }
+}
